requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
eventlet==0.35.2
SQLAlchemy==2.0.34
apscheduler==3.10.4
python-dotenv==1.0.1
requests==2.32.3
backports.zoneinfo==0.2.1;python_version<'3.9'
